+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 ### Added
 - Add an optional indicator for visualizing the paths with different kinds of lighting ([#327](https://gitlab.heigit.org/climate-action/plugins/walkability/-/work_items/327))
 
+### Fixed
+- Catch 413 error and also handle unknown ohsome
+  errors ([#353](https://gitlab.heigit.org/climate-action/plugins/walkability/-/work_items/353))
+
 ## [4.0.3](https://gitlab.heigit.org/climate-action/plugins/walkability/-/releases/4.0.3) - 2026-07-02
 
 ### Changed